adds superclass for enigma machine class
